+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Changes
+- Allow echo route to support GET methods by moving it into it's path (/frontend/echo)
+
 ## 2.1.2 - 2018-11-22
 ### Fixes
 - Fixes typos in README.md